add feat: Auth API

+++ src/pages/api/[[...user]].tsx
@@ -1,5 +1,6 @@
 import { deleteData, retrieveData, updateData } from "@/lib/firebase/service";
 import type { NextApiRequest, NextApiResponse } from "next";
+import jwt from "jsonwebtoken";
 
 export default async function handler(
   req: NextApiRequest,
@@ -37,20 +38,35 @@ export default async function handler(
     });
   } else if (req.method === "DELETE") {
     const { user }: any = req.query;
-    await deleteData("users", user[1], (result: boolean) => {
-      if (result) {
-        res.status(200).json({
-          status: "success",
-          statuscode: 200,
-          message: "success",
-        });
-      } else {
-        res.status(400).json({
-          status: "false",
-          statuscode: 400,
-          message: "failed",
-        });
+    const token = req.headers.authorization?.split(" ")[1] || "";
+    jwt.verify(
+      token,
+      process.env.NEXTAUTH_SECRET || "",
+      async (err: any, decoded: any) => {
+        if (decoded && decoded.role === "admin") {
+          await deleteData("users", user[1], (result: boolean) => {
+            if (result) {
+              res.status(200).json({
+                status: "success",
+                statuscode: 200,
+                message: "success",
+              });
+            } else {
+              res.status(400).json({
+                status: "false",
+                statuscode: 400,
+                message: "failed",
+              });
+            }
+          });
+        } else {
+          res.status(403).json({
+            status: "false",
+            statuscode: 403,
+            message: "Access denied",
+          });
+        }
       }
-    });
+    );
   }
 }
